First upload of MAD functions and kernels
This is the first upload of the MAD functions in relation to the
acceptance of the related paper:
S. Trevisani and M. Rocca. MAD: robust image texture analysis for
applications in high resolution geomorphometry. Computer & Geosciences
(2015), 10.1016/j.cageo.2015.04.003.
For any Issue contact the author: strevisani@iuav.it or
sebastiano.trevisani@gmail.com